+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `redstoneSide` config option can now be a table, if you wish to have multiple outputs for redstone heartbeat
+- It is now possible to use redstone integrators for rs heartbeat using `redstoneIntegrator` option, allows for multiple using same syntax as `redstoneSide`
+- Added `textScale` option to config, which allows monitor scale to be configurable
 
 ## 1.0.1
 ### Fixed